Remove EpoClient callback and move callback to responseCallback
This commit includes updates per underlying changes to the EpoClient API
in the opendxl-epo-client-javascript package. This commit removes the
callback parameter from the EpoClient constructor, since it no longer
exists. This commit also moves the callback parameter to the
runCommand() method to be under the `responseCallback` property in the
`options` object.