Set example params in change nodes
Previously, the parameters used in ePO requests in the examples could be
set across multiple nodes within the flow - the starting inject node, a
change node, and/or the request node itself. In this commit, the
examples have been updated to uniformly set all of the available ePO
request parameters from within the change node which immediately
precedes the ePO request node in the flow.